Refactor env var dict merging into env var storage module
